# be/docs/swagger.yaml
basePath: /
definitions:
...
  api.ErrorResponse:
    properties:
      code:
        description: Error code for frontend i18n
        type: string
      message:
        description: Fallback message in English
        type: string
      params:
        additionalProperties:
          type: string
        description: Dynamic parameters for interpolation
        type: object
    type: object
...
host: localhost:1971
info:
  contact:
    email: support@example.com
    name: API Support
  description: API documentation for R-PRJ-NG backend
  license:
    name: Apache 2.0
    url: https://opensource.org/licenses/Apache-2.0
  termsOfService: http://swagger.io/terms/
  title: ρBee (rhobee) API
  version: "1.0"
paths:
...
  /groups:
    get:
      description: Retrieve a list of all groups
      parameters:
      - description: Search term to filter groups by name
        in: query
        name: search
        type: string
      - description: Field to order the results by
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get all groups
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Create a new group with the provided details
      parameters:
      - description: Group details
        in: body
        name: group
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Group Already Exists
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create a new group
      tags:
      - groups
  /groups/{id}:
    delete:
      description: Delete a group by its ID
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Delete group by ID
      tags:
      - groups
    get:
      description: Retrieve a group by its ID
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Group Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get group by ID
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Update a group's details by its ID
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Group details
        in: body
        name: group
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Group Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Update group by ID
      tags:
      - groups
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...